added support for multiple interfaces
The first interface on the command line is now the trunk,
that's the only required argument, as we expect to be able
to dynamically add "ports" (device interfaces) soon.

Some unit tests were also added (see the tests/ directory).

NOTE ON COMPILATION TIME:

tests/cacth.cpp will take a while to compile (up to a few
minutes on the PI), this is a known limitation of the Catch2
unit test framework.

However, this file should only be compiled once, so subsequent
builds will run faster.

Also, consider installing ccache if you haven't already.